moved constants from types file in react-hooks, page-broker, page-coretime
