+++ packages/page-coretime/src/constants.ts
@@ -1,0 +1,8 @@
+// Copyright 2017-2024 @polkadot/app-coretime authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export const PhaseName = {
+  FixedPrice: 'Fixed Price',
+  PriceDiscovery: 'Price Discovery',
+  Renewals: 'Renewals'
+} as const;
