+++ packages/react-hooks/src/constants.ts
@@ -1,0 +1,28 @@
+// Copyright 2017-2024 @polkadot/react-hooks authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export const AddressIdentityOtherDiscordKey = 'Discord';
+
+export enum CoreTimeTypes {
+  'Reservation',
+  'Lease',
+  'Bulk Coretime',
+  'On Demand'
+}
+
+export const ChainRenewalStatus = {
+  Eligible: 'eligible',
+  None: '-',
+  Renewed: 'renewed'
+};
+
+// RelayChain
+export const CoreTimeConsts = {
+  BlockTime: 6000,
+  BlocksPerTimeslice: 80
+};
+
+export const CoreTimeChainConsts = {
+  BlockTime: 12000,
+  BlocksPerTimeslice: 40
+};
